feat: rss to db

+++ src/processors/database/entities/base.entity.ts
@@ -3,7 +3,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm'
 
 @Entity()
 export class BaseEntity {
-  @PrimaryGeneratedColumn({ type: 'int' })
+  @PrimaryGeneratedColumn({ type: 'bigint' })
   @ApiHideProperty()
   id: string
   @Column({ type: 'datetime', default: () => 'CURRENT_TIMESTAMP' }) // 需要新版 SQL 支持
